refactor: updated useUnsavedChanges -> useUnsavedChangesDialog
